Move test class to own file
