+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/ProjectSystem/VS/PackageRestore/PackageRestoreDataSourceTests.cs
@@ -4,119 +4,9 @@ using Microsoft.VisualStudio.IO;
 using Microsoft.VisualStudio.ProjectSystem.PackageRestore;
 using Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore.Snapshots;
 using Microsoft.VisualStudio.Text;
-using NuGet.SolutionRestoreManager;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
 {
-    public class NuGetRestoreServiceTests
-    {
-        [Fact]
-        public async Task NominateAsyncCallsThroughToNuGetNominate()
-        {
-            bool nominateCalled = false;
-
-            var project = UnconfiguredProjectFactory.Create(fullPath: @"C:\Test\Test.csproj");
-            var vsNuGetSolutionRestoreService = IVsSolutionRestoreServiceFactory.ImplementNominateProjectAsync((path, info, ct) => nominateCalled = true);
-            var vsNuGetSolutionRestoreService4 = IVsSolutionRestoreService4Factory.Create();
-            var projectAsynchronousTasksService = IProjectAsynchronousTasksServiceFactory.Create();
-            var faultHandlerService = IProjectFaultHandlerServiceFactory.Create();
-            var restoreService = new NuGetRestoreService(project, vsNuGetSolutionRestoreService, vsNuGetSolutionRestoreService4, projectAsynchronousTasksService, faultHandlerService);
-
-            var restoreInfo = ProjectRestoreInfoFactory.Create(msbuildProjectExtensionsPath: @"C:\Alpha\Beta");
-            var configuredInputs = PackageRestoreConfiguredInputFactory.Create(restoreInfo);
-
-            var result = await restoreService.NominateAsync(restoreInfo, configuredInputs, default);
-
-            Assert.True(nominateCalled);
-        }
-
-        [Fact]
-        public async Task UpdateDoesNotCallThroughToNuGetNominate()
-        {
-            bool nominateCalled = false;
-
-            var project = UnconfiguredProjectFactory.Create(fullPath: @"C:\Test\Test.csproj");
-            var vsNuGetSolutionRestoreService = IVsSolutionRestoreServiceFactory.ImplementNominateProjectAsync((path, info, ct) => nominateCalled = true);
-            var vsNuGetSolutionRestoreService4 = IVsSolutionRestoreService4Factory.Create();
-            var projectAsynchronousTasksService = IProjectAsynchronousTasksServiceFactory.Create();
-            var faultHandlerService = IProjectFaultHandlerServiceFactory.Create();
-            var restoreService = new NuGetRestoreService(project, vsNuGetSolutionRestoreService, vsNuGetSolutionRestoreService4, projectAsynchronousTasksService, faultHandlerService);
-
-            var restoreInfo = ProjectRestoreInfoFactory.Create(msbuildProjectExtensionsPath: @"C:\Alpha\Beta");
-            var configuredInputs = PackageRestoreConfiguredInputFactory.Create(restoreInfo);
-
-            await restoreService.UpdateWithoutNominationAsync(configuredInputs);
-
-            Assert.False(nominateCalled);
-        }
-
-        [Fact]
-        public async Task NominateCausesPendingTaskToComplete()
-        {
-            IVsProjectRestoreInfoSource? restoreSource = null;
-            Task? faultHandlerRegisteredTask = null;
-            
-            var configuredProject = ConfiguredProjectFactory.Create(projectConfiguration: ProjectConfigurationFactory.Create("Debug|x64"));
-            var project = UnconfiguredProjectFactory.Create(fullPath: @"C:\Test\Test.csproj", configuredProject: configuredProject);
-            var vsNuGetSolutionRestoreService = IVsSolutionRestoreServiceFactory.Create();
-            var vsNuGetSolutionRestoreService4 = IVsSolutionRestoreService4Factory.ImplementRegisterRestoreInfoSourceAsync((source, ct) => restoreSource = source);
-            var projectAsynchronousTasksService = IProjectAsynchronousTasksServiceFactory.Create();
-            var faultHandlerService = IProjectFaultHandlerServiceFactory.ImplementForget((task, settings, severity, project) => faultHandlerRegisteredTask = task);
-            var restoreService = new NuGetRestoreService(project, vsNuGetSolutionRestoreService, vsNuGetSolutionRestoreService4, projectAsynchronousTasksService, faultHandlerService);
-            
-            await restoreService.LoadAsync();
-
-            Assert.NotNull(faultHandlerRegisteredTask);
-
-            await faultHandlerRegisteredTask;
-
-            Assert.NotNull(restoreSource);
-
-            Task nominationTask = restoreSource.WhenNominated(default);
-            Assert.False(nominationTask.IsCompleted);
-
-            var restoreInfo = ProjectRestoreInfoFactory.Create(msbuildProjectExtensionsPath: @"C:\Alpha\Beta");
-            var configuredInputs = PackageRestoreConfiguredInputFactory.Create(restoreInfo);
-
-            await restoreService.NominateAsync(restoreInfo, configuredInputs, default);
-
-            Assert.True(nominationTask.IsCompleted);
-        }
-
-        [Fact]
-        public async Task UpdateCausesPendingTaskToComplete()
-        {
-            IVsProjectRestoreInfoSource? restoreSource = null;
-            Task? faultHandlerRegisteredTask = null;
-
-            var configuredProject = ConfiguredProjectFactory.Create(projectConfiguration: ProjectConfigurationFactory.Create("Debug|x64"));
-            var project = UnconfiguredProjectFactory.Create(fullPath: @"C:\Test\Test.csproj", configuredProject: configuredProject);
-            var vsNuGetSolutionRestoreService = IVsSolutionRestoreServiceFactory.Create();
-            var vsNuGetSolutionRestoreService4 = IVsSolutionRestoreService4Factory.ImplementRegisterRestoreInfoSourceAsync((source, ct) => restoreSource = source);
-            var projectAsynchronousTasksService = IProjectAsynchronousTasksServiceFactory.Create();
-            var faultHandlerService = IProjectFaultHandlerServiceFactory.ImplementForget((task, settings, severity, project) => faultHandlerRegisteredTask = task);
-            var restoreService = new NuGetRestoreService(project, vsNuGetSolutionRestoreService, vsNuGetSolutionRestoreService4, projectAsynchronousTasksService, faultHandlerService);
-
-            await restoreService.LoadAsync();
-
-            Assert.NotNull(faultHandlerRegisteredTask);
-
-            await faultHandlerRegisteredTask;
-
-            Assert.NotNull(restoreSource);
-
-            Task nominationTask = restoreSource.WhenNominated(default);
-            Assert.False(nominationTask.IsCompleted);
-
-            var restoreInfo = ProjectRestoreInfoFactory.Create(msbuildProjectExtensionsPath: @"C:\Alpha\Beta");
-            var configuredInputs = PackageRestoreConfiguredInputFactory.Create(restoreInfo);
-
-            await restoreService.UpdateWithoutNominationAsync(configuredInputs);
-
-            Assert.True(nominationTask.IsCompleted);
-        }
-    }
-
     public class PackageRestoreDataSourceTests
     {
         [Fact]
